requirements.txt
PyYAML
requests
requests_oauthlib